+++ CHANGELOG.md
@@ -29,7 +29,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * Missing comma in gpkg_2d_gridded_coverage_ancillary table schema
 * gpkg_data_columns table: removed foreign key requirement, unique constraint update
 * Missing comma in nga_tile_scaling table schema
-* OGC OpenAPI Features download to GeoPackage support
+* OGC API Features download to GeoPackage support
 
 ## [3.2.0](https://github.com/ngageoint/geopackage-core-java/releases/tag/3.2.0) (04-02-2019)